cleanup. reduced content of the engine methods that were moved into the Adapter project.
