+++ Etabs_Adapter/Read/Node.cs
@@ -81,7 +81,7 @@ namespace BH.Adapter.ETABS
                 m_model.PointObj.GetRestraint(id, ref restraint);
                 m_model.PointObj.GetSpring(id, ref spring);
 
-                Constraint6DOF support = Engine.ETABS.Convert.GetConstraint6DOF(restraint, spring);
+                Constraint6DOF support = GetConstraint6DOF(restraint, spring);
 
                 Node bhNode = Engine.Structure.Create.Node(new oM.Geometry.Point() { X = x, Y = y, Z = z }, "", support);
                 bhNode.CustomData.Add(AdapterId, id);
@@ -91,6 +91,28 @@ namespace BH.Adapter.ETABS
 
 
             return nodeList;
+        }
+
+        /***************************************************/
+
+        public static Constraint6DOF GetConstraint6DOF(bool[] restraint, double[] springs)
+        {
+            Constraint6DOF bhConstraint = new Constraint6DOF();
+            bhConstraint.TranslationX = restraint[0] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.TranslationY = restraint[1] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.TranslationZ = restraint[2] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationX = restraint[3] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationY = restraint[4] == true ? DOFType.Fixed : DOFType.Free;
+            bhConstraint.RotationZ = restraint[5] == true ? DOFType.Fixed : DOFType.Free;
+
+            bhConstraint.TranslationalStiffnessX = springs[0];
+            bhConstraint.TranslationalStiffnessY = springs[1];
+            bhConstraint.TranslationalStiffnessZ = springs[2];
+            bhConstraint.RotationalStiffnessX = springs[3];
+            bhConstraint.RotationalStiffnessY = springs[4];
+            bhConstraint.RotationalStiffnessZ = springs[5];
+
+            return bhConstraint;
         }
 
         /***************************************************/
