+++ Etabs_Adapter/_Engine/Convert/ToBHoM.cs
@@ -47,57 +47,7 @@ namespace BH.Engine.ETABS
 {
     public static partial class Convert
     {
-        /***************************************************/
-
-        public static List<Node> ToBHoM(this cPointObj pointObj, List<string> ids)
-        {
-            List<Node> bhNodes = new List<Node>();
-            int nameCount = 0;
-            string[] nameArr = { };
-
-            if (ids == null)
-            {
-                pointObj.GetNameList(ref nameCount, ref nameArr);
-                ids = nameArr.ToList();
-            }
-
-            foreach (string id in ids)
-            {
-                bhNodes.Add(pointObj.ToBHoM(id));
-            }
-
-            return bhNodes;
-        }
-
-        /***************************************************/
-
-        public static Node ToBHoM(this cPointObj pointObj, string id)
-        {
-
-            double x, y, z;
-            x = y = z = 0;
-            bool[] restraint = new bool[6];
-            double[] spring = new double[6];
-
-            pointObj.GetCoordCartesian(id, ref x, ref y, ref z);
-
-            pointObj.GetRestraint(id, ref restraint);
-            pointObj.SetSpring(id, ref spring);
-
-            Node bhNode = Structure.Create.Node(new oM.Geometry.Point() { X = x, Y = y, Z = z }, "", GetConstraint6DOF(restraint, spring));
-
-            string AdapterId;
-
-#if Debug2017
-            AdapterId = ETABS2017Adapter.ID;
-#else
-            AdapterId = ETABS2016Adapter.ID;
-#endif
-            bhNode.CustomData.Add(AdapterId, id);
-
-            return bhNode;
-        }
-
+        
         /***************************************************/
 
         public static LoadNature ToBHoM(this eLoadPatternType loadPatternType)
@@ -170,69 +120,6 @@ namespace BH.Engine.ETABS
                 default:
                     return "None";
             }
-        }
-
-        /***************************************************/
-
-        public static Constraint6DOF GetConstraint6DOF(bool[] restraint, double[] springs)
-        {
-            Constraint6DOF bhConstraint = new Constraint6DOF();
-            bhConstraint.TranslationX = restraint[0] == true ? DOFType.Fixed : DOFType.Free;
-            bhConstraint.TranslationY = restraint[1] == true ? DOFType.Fixed : DOFType.Free;
-            bhConstraint.TranslationZ = restraint[2] == true ? DOFType.Fixed : DOFType.Free;
-            bhConstraint.RotationX = restraint[3] == true ? DOFType.Fixed : DOFType.Free;
-            bhConstraint.RotationY = restraint[4] == true ? DOFType.Fixed : DOFType.Free;
-            bhConstraint.RotationZ = restraint[5] == true ? DOFType.Fixed : DOFType.Free;
-
-            bhConstraint.TranslationalStiffnessX = springs[0];
-            bhConstraint.TranslationalStiffnessY = springs[1];
-            bhConstraint.TranslationalStiffnessZ = springs[2];
-            bhConstraint.RotationalStiffnessX = springs[3];
-            bhConstraint.RotationalStiffnessY = springs[4];
-            bhConstraint.RotationalStiffnessZ = springs[5];
-
-            return bhConstraint;
-        }
-
-        /***************************************************/
-
-        public static BarRelease GetBarRelease(bool[] startRestraint, double[] startSpring, bool[] endRestraint, double[] endSpring)
-        {
-            Constraint6DOF startRelease = new Constraint6DOF();
-
-            startRelease.TranslationX = startRestraint[0] == true ? DOFType.Free : DOFType.Fixed;
-            startRelease.TranslationY = startRestraint[1] == true ? DOFType.Free : DOFType.Fixed;
-            startRelease.TranslationZ = startRestraint[2] == true ? DOFType.Free : DOFType.Fixed;
-            startRelease.RotationX = startRestraint[3] == true ? DOFType.Free : DOFType.Fixed;
-            startRelease.RotationY = startRestraint[4] == true ? DOFType.Free : DOFType.Fixed;
-            startRelease.RotationZ = startRestraint[5] == true ? DOFType.Free : DOFType.Fixed;
-
-            startRelease.TranslationalStiffnessX = startSpring[0];
-            startRelease.TranslationalStiffnessY = startSpring[1];
-            startRelease.TranslationalStiffnessZ = startSpring[2];
-            startRelease.RotationalStiffnessX = startSpring[3];
-            startRelease.RotationalStiffnessY = startSpring[4];
-            startRelease.RotationalStiffnessZ = startSpring[5];
-
-            Constraint6DOF endRelease = new Constraint6DOF();
-
-            endRelease.TranslationX = endRestraint[0] == true ? DOFType.Free : DOFType.Fixed;
-            endRelease.TranslationY = endRestraint[1] == true ? DOFType.Free : DOFType.Fixed;
-            endRelease.TranslationZ = endRestraint[2] == true ? DOFType.Free : DOFType.Fixed;
-            endRelease.RotationX = endRestraint[3] == true ? DOFType.Free : DOFType.Fixed;
-            endRelease.RotationY = endRestraint[4] == true ? DOFType.Free : DOFType.Fixed;
-            endRelease.RotationZ = endRestraint[5] == true ? DOFType.Free : DOFType.Fixed;
-
-            endRelease.TranslationalStiffnessX = endSpring[0];
-            endRelease.TranslationalStiffnessY = endSpring[1];
-            endRelease.TranslationalStiffnessZ = endSpring[2];
-            endRelease.RotationalStiffnessX = endSpring[3];
-            endRelease.RotationalStiffnessY = endSpring[4];
-            endRelease.RotationalStiffnessZ = endSpring[5];
-
-            BarRelease barRelease = new BarRelease() { StartRelease = startRelease, EndRelease = endRelease };
-
-            return barRelease;
         }
     }
 }
